Renames wunderground -> redemet
This basically renames the module. The internal names will change later
and the database table will probably keep the `Clima_wu` name for
backwards compatibility.
