+++ crawlclima/pegatemperatura.py
@@ -4,11 +4,14 @@ Este script deve ser executado via cron, diariamente para
 lançar a tarefa de captura de dados de temperatura do Weather underground
 veja crontab para saber como executar este script
 """
-import sys, os
+import sys
+import os
 from datetime import datetime, timedelta
 sys.path.append(os.getcwd())
-from crawlclima.tasks import fetch_wunderground
-from utilities.models import save, find_all
+
+from crawlclima.tasks import fetch_redemet
+from utilities.models import find_all
+
 
 user_date = None
 if len(sys.argv) > 1:
@@ -26,4 +29,4 @@ stations = [row['estacao_id'] for row in rows]
 day = year_start if today.isoweekday() == 5 else yesterday
 
 for station in stations:
-    fetch_wunderground.delay(station, day)
+    fetch_redemet.delay(station, day)
