+++ crawlclima/redemet/tests/test_rmet.py
@@ -3,7 +3,7 @@
 from datetime import datetime
 import unittest
 import pandas as pd
-from crawlclima.wunderground.wu import parse_page, fahrenheit_to_celsius, wu_url, date_generator, \
+from crawlclima.redemet.rmet import parse_page, fahrenheit_to_celsius, wu_url, date_generator, \
     describe, capture_date_range, capture
 
 
@@ -98,13 +98,13 @@ class TestDateGenerator(unittest.TestCase):
 
 class TestParsePage(unittest.TestCase):
     def testCelsiusDailyHistory(self):
-        with open('crawlclima/wunderground/tests/CelsiusDailyHistory.html', 'r') as fd:
+        with open('crawlclima/redemet/tests/CelsiusDailyHistory.html', 'r') as fd:
             dataframe = parse_page(fd.read())
         self.assertEqual(dataframe.DateUTC[0], '2015-08-01 03:00:00')
         self.assertAlmostEqual(dataframe.TemperatureC.mean(), 24.33, 2)
 
     def testEmptyDailyHistory(self):
-        with open('crawlclima/wunderground/tests/EmptyDailyHistory.html', 'r') as fd:
+        with open('crawlclima/redemet/tests/EmptyDailyHistory.html', 'r') as fd:
             dataframe = parse_page(fd.read())
 
         self.assertIsInstance(dataframe, pd.DataFrame)
@@ -117,7 +117,7 @@ class TestParsePage(unittest.TestCase):
 
 class TestDescribe(unittest.TestCase):
     def test_filled_dataframe(self):
-        with open('crawlclima/wunderground/tests/CelsiusDailyHistory.html', 'r') as fd:
+        with open('crawlclima/redemet/tests/CelsiusDailyHistory.html', 'r') as fd:
             dataframe = parse_page(fd.read())
 
         summary = describe(dataframe)
@@ -134,7 +134,7 @@ class TestDescribe(unittest.TestCase):
         })
 
     def test_empty_dataframe(self):
-        with open('crawlclima/wunderground/tests/EmptyDailyHistory.html', 'r') as fd:
+        with open('crawlclima/redemet/tests/EmptyDailyHistory.html', 'r') as fd:
             dataframe = parse_page(fd.read())
 
         summary = describe(dataframe)
